proyecto final/requirements.txt
psycopg2-binary==2.9.7
reportlab==4.0.4